logic-aarch64: Fix extractnarrow() when src and dst are the same
If the source and destination are the same, then the top part of the
vector register will be set to all zero before the narrowing operation
takes place.

The solution is to clear the top part of the register after the
narrowing operation has taken place, so that the vector operation still
has access to the data in the upper lanes.

This also mimics the ARM ARM's pseudocode a little more closely.